unify the merge results class
